+++ clang/docs/ReleaseNotes.rst
@@ -306,6 +306,7 @@ Python Binding Changes
 
 OpenMP Support
 --------------
+- Added support for 'omp assume' directive.
 
 Additional Information
 ======================